+++ visualization/app/codeCharta/state/store/appSettings/invertHeight/invertHeight.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { appSettingsSelector } from "../appSettings.selector"
+
+export const invertHeightSelector = createSelector([appSettingsSelector], appSettings => appSettings.invertHeight)