+++ visualization/app/codeCharta/ui/slider/slider.component.ts
@@ -10,7 +10,9 @@ export class SliderComponent {
 	@Input() value?: number
 	@Input() min: number
 	@Input() max: number
+	@Input() label: string
 	@Input() step?: number = 1
+	@Input() disabled?: boolean = false
 	@Input() onChange: (number) => void
 
 	handleSliderOnChange($event: MatSliderChange) {